+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/servlet/taglib/HorizontalCard.java
@@ -12,42 +12,17 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.servlet.taglib.soy;
+package com.liferay.frontend.taglib.clay.servlet.taglib;
 
 /**
  * @author Eudaldo Alonso
  */
-public interface UserCard extends BaseClayCard {
+public interface HorizontalCard extends BaseClayCard {
 
 	public default String getIcon() {
 		return null;
 	}
 
-	public default String getImageAlt() {
-		return null;
-	}
-
-	public default String getImageSrc() {
-		return null;
-	}
-
-	public String getName();
-
-	public default String getSubtitle() {
-		return null;
-	}
-
-	public default String getUserColorClass() {
-		return null;
-	}
-
-	/**
-	 * @deprecated As of Athanasius (7.3.x), replaced by {@link
-	 *             #getUserColorClass()}
-	 */
-	@Deprecated
-	public default String getUserColorCssClass() {
-		return null;
-	}
+	public String getTitle();
 
 }
